backend/requirements.txt
django
djangorestframework
django-cors-headers
pillow
python-dotenv
djangorestframework-simplejwt
django-recaptcha
django-allauth